fixed the version bug on PPC architecture

+++ src/cairo-dock-modules.c
@@ -59,7 +59,7 @@ extern CairoDock *g_pMainDock;
 extern gchar *g_cConfFile;
 extern gchar *g_cCurrentThemePath;
 extern gchar *g_cCairoDockDataDir;
-extern short g_iMajorVersion, g_iMinorVersion, g_iMicroVersion;
+extern int g_iMajorVersion, g_iMinorVersion, g_iMicroVersion;
 extern gboolean g_bEasterEggs;
 
 static GHashTable *s_hModuleTable = NULL;
